Move the changes to new file
